Updating configure.in, see kom-rsvp rev 72, /branches/jason_force10:
---
Added some autoconf magic to configure.in to check for the crypto
libs.  If they cannot be found by configure, try compiling kom-rsvp
w/o -lcrypto in LDFLAGS.  This should help the build process complete
successfully if net-snmp was not compiled with the crypto libs.
---

These configure options have _not_ yet been migrated to configure.in:

   --with-switch-ctrl-port
   --with-force10-software-v6
   --with-switch-vendor-model
   --enable-switch-cli-access
   --enable-switch-cli-type




git-svn-id: file:///var/www/svn/dragon-sw/trunk@868 625cd551-afe4-0310-9ad4-9d199310cf85