A more generic solution for the below fix:
Fixed a bug: Ratptor swithes allows for a port to be untagged in multiple VLANs. Since we used findVlanByUntagged port to look for the VLAN we previously moved the ports into, this may confuse the logic.
Solution: used a private variable activeVlanId in class SwitchCtrl_Session_RaptorER1010 to track the VLAN being operated.

git-svn-id: file:///var/www/svn/dragon-sw/trunk@871 625cd551-afe4-0310-9ad4-9d199310cf85