Committing Ajay's modifications to the VLSR code to include support for the Cisco Catalyst switches exclusively via SNMP.
We have tested with the Cisco Catalyst 3750 but further testing still needs to be done with the 65xx
series switches.

Testing with the Cisco 3750 was done with IOS 12.2(37)SE.  Older IOS revisions may have problems,
as we started with 12.1(14)EA and had MIB-related issues...(some newer MIBs are not supported).

There is not currently any support for bandwidth policing/limiting yet.

Several small, unrelated fixed to configure/configure.in are also being committed at this time.
These fixes allow autoconf to re-generate the configure script without any issues.

The 'do_build.sh' script needs to be updated to include the following new target:
  vlsr-cat3750|vlsr-catalyst3750|VLSR-CATALYST3750|VLSR-Catalyst3750)
    rsvpconf=`echo "--with-switch-vendor-model=Catalyst3750 --with-switch-ctrl-port=255 --disable-altq --enable-switch-port-shutdown"`
  ;;

The Cisco 3750 switch was using the following image file during our tests:
  Cisco IOS Software, C3750 Software (C3750-IPSERVICESK9-M), Version 12.2(37)SE, RELEASE SOFTWARE (fc2)
  System image file is "flash:c3750-ipservicesk9-mz.122-37.SE.bin"



git-svn-id: file:///var/www/svn/dragon-sw/trunk@1522 625cd551-afe4-0310-9ad4-9d199310cf85