Support for the following described scenario
On an oc-192 link, when one STS1 has been taken, 
OSRP will consider oc-48 the maximum reservable
bandwidth. Therefore no single SNC over oc-48 can
be created. As a solution for this scenario, we
need to create up to 4 SNCs. The implication is
that up to 4 GTPs should be created (and CRS's 
in the case we need local cross-connects).


git-svn-id: file:///var/www/svn/dragon-sw/trunk@1533 625cd551-afe4-0310-9ad4-9d199310cf85