Small fix to kom-rsvp's configure/configure.in.
When --enable-switch-cli-access option is present, the text printed to
the user was somewhat misleading.  The user is told that the choices
are "TELNET, SSH, TL1 or SHELL" but we are really expecting the user
to enter their response in lowercase.

The options have been switched to lowercase w/ single quotes around
them.  The 'if test ...' lines now check for both all-uppercase and
all-lowercase input from the user, but unfortunately not a mixture of
both.  If the user enters 'Tl1' or 'Telnet' it will not be interpreted
correctly...(perhaps we should add a while statement so that the user
is forced to enter a valid response...)



git-svn-id: file:///var/www/svn/dragon-sw/trunk@1711 625cd551-afe4-0310-9ad4-9d199310cf85