Addition of new switch module for SMC 10G 8708 switch (an 8-port 10GigE switch) from code contributed by Yujiang (John) Qu <Yujiang.Qu@ngc.com>.
Things to note regarding this code:

* the SMC 8707 is RFC2674 compliant and this code *does* seem to create VLANs 
dynamically, but I noticed that all the code under removeVLAN is commented out.  
There is a comment under removeVLAN() that simply says "Keep empty vlan on switch
for Now", meaning that the ports are removed from the VLAN but the VLAN itself is
not removed.

* they *do* seem to have found a way to do policing using the SMC private MIB, 
see the readPortIngressRateListFromSwitch() function, they are using this OID
from the SMC private MIB: .1.3.6.1.4.1.202.20.47.1.16.1.2.1.2.

* there are a few comments in the code regarding setting of the PVID on these
switches

* the SMC 8708 code was originally written in May 2007, so it needs to be tested
with the other changes that have been made to the SwitchCtrl code (in particular,
will revisions 1725-1733 affect it?)

* we should make a note of the software rev on the SMC 8708 that was used for 
testing




git-svn-id: file:///var/www/svn/dragon-sw/trunk@1737 625cd551-afe4-0310-9ad4-9d199310cf85