Modifications to kom-rsvp to support compilation under Darwin (Mac OS X).
Also updated configure.in/configure script to eliminate this compile warning:
(took out the MAKEFLAGS stuff)

''warning: -jN forced in submake: disabling jobserver mode...''

NOTE: Although kom-rsvp now compiles under Darwin, there is still a run-time
issue when gathering the system interfaces that we are investigating.  It is
not *exactly* the same as FreeBSD, but it is very similar.




git-svn-id: file:///var/www/svn/dragon-sw/trunk@1847 625cd551-afe4-0310-9ad4-9d199310cf85