Remove -lcrypt from ast_master_LDADD and node_agent_LDADD because it is unnecessary.
autoconf already checks for crypt() in -lcrypt and will stick the
right thing into LIBS as neeeded.  (Just like it does when building
the zebra/dragon/ directory...)

Having -lcrypt in those LDADD's breaks compilation under Darwin (Mac
OS X) because -lcrypt cannot be used (and is unnecessary).  For more
information, please refer to this page and search for '-lcrypt':
 http://developer.apple.com/technotes/tn2002/tn2071.html



git-svn-id: file:///var/www/svn/dragon-sw/trunk@1849 625cd551-afe4-0310-9ad4-9d199310cf85