Fixes to parser.l:
* Change pattern match for comments, allow comments to start with either
  '#' or ';' and also allow lines to contain *only* the comment character
  with nothing following it.

Update parser.y and RSVPD.cc to also log errors to RSVPD.log instead of just
using 'cerr'.  The output from 'cerr' will never be seen by the user when RSVPD
is running in daemon mode with the '-d' flag.  This will at least make it clear
if there is a problem with RSVPD.conf in the log file.



git-svn-id: file:///var/www/svn/dragon-sw/trunk@2061 625cd551-afe4-0310-9ad4-9d199310cf85