Changes to kom-rsvp configure.in and MakeConfig.in necessary to check for libxml2 and add the appropriate -I... -L... -lxml2 directives when compiling.
Also re-ran 'autoconf -I autoconf_in' in the kom-rsvp directory after updating configure.in.

NOTE:  We could avoid the check for net-snmp in do_build.sh (e.g. 
'determine a valid path for net-snmp header files') by modifying
kom-rsvp/configure.in in a similar manner for the --with-snmp option...
(So that configure automatically finds whether netsnmp is installed
in /usr or /usr/local and does not rely on the do_build.sh script
to pass the right option to --with-snmp...)



git-svn-id: file:///var/www/svn/dragon-sw/trunk@2161 625cd551-afe4-0310-9ad4-9d199310cf85