encountered some strange issues with these changes related to the netsnmp library...
the old build (r2161) seemed to prefer the netsnmp in /usr/local/lib, while this
new version prefers the netsnmp in /usr/lib.  These are not always the same, and 
in our case, didn't work at all when the latter was preferred...

Backing out changes until we do further testing on this.





git-svn-id: file:///var/www/svn/dragon-sw/trunk@2178 625cd551-afe4-0310-9ad4-9d199310cf85