Fixed a potential trouble causer: The code forbade provisioning a port if it has been included in another VLAN (VID>1) as UNTAGGED port, but it allowed a port being provisioned into another VLAN when it has been in another VLAN as TAGGED port. While this is legitimate for some switches that can provision a port as untagged and tagged simultanuously, switches such as cisco catalysts will move the tagged port out of the original vlan and place it into the new vlan in access mode, which may unintendedly disrupt existing service.
Here is the first step of the solution: forbidding provision a port when it has be included in another VLAN (VID>1) no matter whether it is tagged or untagged.

Next we will allow users to add options in RSVPD.conf to bypass the whole restriction, or loose the restriction to allow overriding untagged port by including it into another VLAN as tagged or vice versa.



git-svn-id: file:///var/www/svn/dragon-sw/trunk@2252 625cd551-afe4-0310-9ad4-9d199310cf85